rename TrackSelector to StreamSelector and introduce TrackUtils
