+++ new-player/src/main/java/net/newpipe/newplayer/model/NewPlayerViewModelImpl.kt
@@ -51,7 +51,6 @@ import net.newpipe.newplayer.utils.NewPlayerException
 import net.newpipe.newplayer.PlayMode
 import net.newpipe.newplayer.RepeatMode
 import net.newpipe.newplayer.ui.ContentScale
-import net.newpipe.newplayer.utils.TrackSelector
 import java.util.LinkedList
 import kotlin.math.abs
